refactor: move handleMap to separate file
